Change validators

+++ src/main/java/controller/service/impl/AdminServiceImpl.java
@@ -1,7 +1,9 @@
 package controller.service.impl;
 
+import controller.exception.InvalidDataRuntimeException;
 import controller.service.AdminService;
 import controller.service.mapper.AdminMapper;
+import controller.service.validator.impl.AdminValidator;
 import domain.Admin;
 import model.dao.AdminDAO;
 import model.entity.AdminEntity;
@@ -9,6 +11,7 @@ import org.apache.log4j.Logger;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class AdminServiceImpl implements AdminService {
@@ -16,15 +19,21 @@ public class AdminServiceImpl implements AdminService {
 
     private AdminDAO adminDAO;
     private AdminMapper mapper;
+    private AdminValidator adminValidator;
 
-    public AdminServiceImpl(AdminDAO adminDAO, AdminMapper mapper) {
+    public AdminServiceImpl(AdminDAO adminDAO, AdminMapper mapper, AdminValidator adminValidator) {
         this.adminDAO = adminDAO;
         this.mapper = mapper;
+        this.adminValidator = adminValidator;
     }
 
     @Override
     public Admin getAdminByUserId(Integer idUser) {
         LOGGER.info("Try to get admin info by user id");
+        if (Objects.isNull(idUser)) {
+            LOGGER.error("Id user is not exist");
+            throw new InvalidDataRuntimeException("Id user is not exist");
+        }
         AdminEntity adminEntity = adminDAO.getAdminByUserId(idUser);
         return mapper.mapAdminEntityToAdmin(adminEntity);
     }
@@ -32,6 +41,11 @@ public class AdminServiceImpl implements AdminService {
     @Override
     public Integer insertElement(Admin element) {
         LOGGER.info("Inserting element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect insert value");
+            throw new InvalidDataRuntimeException("Incorrect insert value");
+        }
+        adminValidator.validate(element);
         AdminEntity adminEntity = mapper.mapAdminToAdminEntity(element);
         return adminDAO.insertElement(adminEntity);
     }
@@ -39,6 +53,10 @@ public class AdminServiceImpl implements AdminService {
     @Override
     public Admin getElementById(Integer id) {
         LOGGER.info("Try to get element by id");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect getElementById value");
+            throw new InvalidDataRuntimeException("Incorrect getElementById value");
+        }
         AdminEntity adminEntity = adminDAO.getAdminByUserId(id);
         return mapper.mapAdminEntityToAdmin(adminEntity);
     }
@@ -46,12 +64,21 @@ public class AdminServiceImpl implements AdminService {
     @Override
     public void deleteElement(Integer id) {
         LOGGER.info("Deleting element");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect deleteElement value");
+            throw new InvalidDataRuntimeException("Incorrect deleteElement value");
+        }
         adminDAO.deleteElement(id);
     }
 
     @Override
     public void updateElement(Admin element) {
         LOGGER.info("Updating element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect updateElement value");
+            throw new InvalidDataRuntimeException("Incorrect updateElement value");
+        }
+        adminValidator.validate(element);
         AdminEntity adminEntity = mapper.mapAdminToAdminEntity(element);
         adminDAO.updateElement(adminEntity);
     }
