+++ src/main/java/controller/service/impl/BusServiceImpl.java
@@ -1,7 +1,9 @@
 package controller.service.impl;
 
+import controller.exception.InvalidDataRuntimeException;
 import controller.service.BusService;
 import controller.service.mapper.BusMapper;
+import controller.service.validator.impl.BusValidator;
 import domain.Bus;
 import model.dao.BusDAO;
 import model.entity.BusEntity;
@@ -9,6 +11,7 @@ import org.apache.log4j.Logger;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class BusServiceImpl implements BusService {
@@ -16,10 +19,12 @@ public class BusServiceImpl implements BusService {
 
     private BusDAO busDAO;
     private BusMapper mapper;
+    private BusValidator busValidator;
 
-    public BusServiceImpl(BusDAO busDAO, BusMapper mapper) {
+    public BusServiceImpl(BusDAO busDAO, BusMapper mapper, BusValidator busValidator) {
         this.busDAO = busDAO;
         this.mapper = mapper;
+        this.busValidator = busValidator;
     }
 
     @Override
@@ -35,24 +40,40 @@ public class BusServiceImpl implements BusService {
     @Override
     public Integer countBusesOnRouteById(Integer idRoute) {
         LOGGER.info("Counting buses on route");
+        if (Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect countBusesOnRouteById value");
+            throw new InvalidDataRuntimeException("Incorrect countBusesOnRouteById value");
+        }
         return busDAO.countBusesOnRouteById(idRoute);
     }
 
     @Override
     public void cancelBusFromRoute(Integer idBus) {
         LOGGER.info("Try cancel bus from route");
+        if (Objects.isNull(idBus)) {
+            LOGGER.error("Incorrect cancelBusFromRoute value");
+            throw new InvalidDataRuntimeException("Incorrect cancelBusFromRoute value");
+        }
         busDAO.cancelBusFromRoute(idBus);
     }
 
     @Override
     public void appointBusToRoute(Integer idRoute, Integer idBus) {
         LOGGER.info("Assigning appoint bus to route");
+        if (Objects.isNull(idBus) || Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect appointBusToRoute value");
+            throw new InvalidDataRuntimeException("Incorrect appointBusToRoute value");
+        }
         busDAO.appointBusToRoute(idRoute, idBus);
     }
 
     @Override
     public List<Bus> getBusesByIdRoute(Integer idRoute) {
         LOGGER.info("Getting buses by route id");
+        if (Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect getBusesByIdRoute value");
+            throw new InvalidDataRuntimeException("Incorrect getBusesByIdRoute value");
+        }
         List<BusEntity> result = busDAO.getBusesByIdRoute(idRoute);
         return result.isEmpty() ? Collections.emptyList()
                 : result.stream()
@@ -63,6 +84,11 @@ public class BusServiceImpl implements BusService {
     @Override
     public Integer insertElement(Bus element) {
         LOGGER.info("Inserting element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect insertElement value");
+            throw new InvalidDataRuntimeException("Incorrect insertElement value");
+        }
+        busValidator.validate(element);
         BusEntity busEntity = mapper.mapBusToBusEntity(element);
         return busDAO.insertElement(busEntity);
     }
@@ -70,6 +96,10 @@ public class BusServiceImpl implements BusService {
     @Override
     public Bus getElementById(Integer id) {
         LOGGER.info("Try to get element by id");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect getElementById value");
+            throw new InvalidDataRuntimeException("Incorrect getElementById value");
+        }
         BusEntity busEntity = busDAO.getElementById(id);
         return mapper.mapBusEntityToBus(busEntity);
     }
@@ -77,12 +107,21 @@ public class BusServiceImpl implements BusService {
     @Override
     public void deleteElement(Integer id) {
         LOGGER.info("Deleting element");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect deleteElement value");
+            throw new InvalidDataRuntimeException("Incorrect deleteElement value");
+        }
         busDAO.deleteElement(id);
     }
 
     @Override
     public void updateElement(Bus element) {
         LOGGER.info("Updating element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect updateElement value");
+            throw new InvalidDataRuntimeException("Incorrect updateElement value");
+        }
+        busValidator.validate(element);
         BusEntity busEntity = mapper.mapBusToBusEntity(element);
         busDAO.updateElement(busEntity);
     }
