+++ src/main/java/controller/service/impl/RouteServiceImpl.java
@@ -1,7 +1,9 @@
 package controller.service.impl;
 
+import controller.exception.InvalidDataRuntimeException;
 import controller.service.RouteService;
 import controller.service.mapper.RouteMapper;
+import controller.service.validator.impl.RouteValidator;
 import domain.Route;
 import model.dao.RouteDAO;
 import model.entity.RouteEntity;
@@ -9,6 +11,7 @@ import org.apache.log4j.Logger;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class RouteServiceImpl implements RouteService {
@@ -16,33 +19,51 @@ public class RouteServiceImpl implements RouteService {
 
     private RouteDAO routeDAO;
     private RouteMapper mapper;
+    private RouteValidator routeValidator;
 
-    public RouteServiceImpl(RouteDAO routeDAO, RouteMapper mapper) {
+    public RouteServiceImpl(RouteDAO routeDAO, RouteMapper mapper, RouteValidator routeValidator) {
         this.routeDAO = routeDAO;
         this.mapper = mapper;
+        this.routeValidator = routeValidator;
     }
 
     @Override
     public void setStatusEmpty(Integer idRoute) {
         LOGGER.info("Setting status empty for route");
+        if (Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect setStatusEmpty value");
+            throw new InvalidDataRuntimeException("Incorrect setStatusEmpty value");
+        }
         routeDAO.setStatusEmpty(idRoute);
     }
 
     @Override
     public void setStatusWork(Integer idRoute) {
         LOGGER.info("Setting status work for route");
+        if (Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect setStatusWork value");
+            throw new InvalidDataRuntimeException("Incorrect setStatusWork value");
+        }
         routeDAO.setStatusWork(idRoute);
     }
 
     @Override
     public void cancelAll(Integer idRoute) {
         LOGGER.info("Cancel all buses from route");
+        if (Objects.isNull(idRoute)) {
+            LOGGER.error("Incorrect cancelAll value");
+            throw new InvalidDataRuntimeException("Incorrect cancelAll value");
+        }
         routeDAO.cancelAll(idRoute);
     }
 
     @Override
     public List<Route> searchByCriteria(String departure, String arrival) {
         LOGGER.info("Searching by criteria");
+        if (departure.isEmpty() || arrival.isEmpty()) {
+            LOGGER.error("Incorrect searchByCriteria value");
+            throw new InvalidDataRuntimeException("Incorrect searchByCriteria value");
+        }
         List<RouteEntity> result = routeDAO.searchByCriteria(departure, arrival);
         return result.isEmpty() ? Collections.emptyList()
                 : result.stream()
@@ -53,6 +74,11 @@ public class RouteServiceImpl implements RouteService {
     @Override
     public Integer insertElement(Route element) {
         LOGGER.info("Inserting element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect insertElement value");
+            throw new InvalidDataRuntimeException("Incorrect insertElement value");
+        }
+        routeValidator.validate(element);
         RouteEntity driverEntity = mapper.mapRouteToRouteEntity(element);
         return routeDAO.insertElement(driverEntity);
     }
@@ -60,6 +86,10 @@ public class RouteServiceImpl implements RouteService {
     @Override
     public Route getElementById(Integer id) {
         LOGGER.info("Try to get element by id");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect getElementById value");
+            throw new InvalidDataRuntimeException("Incorrect getElementById value");
+        }
         RouteEntity routeEntity = routeDAO.getElementById(id);
         return mapper.mapRouteEntityToRoute(routeEntity);
     }
@@ -67,12 +97,21 @@ public class RouteServiceImpl implements RouteService {
     @Override
     public void deleteElement(Integer id) {
         LOGGER.info("Deleting element");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect deleteElement value");
+            throw new InvalidDataRuntimeException("Incorrect deleteElement value");
+        }
         routeDAO.deleteElement(id);
     }
 
     @Override
     public void updateElement(Route element) {
         LOGGER.info("Updating element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect updateElement value");
+            throw new InvalidDataRuntimeException("Incorrect updateElement value");
+        }
+        routeValidator.validate(element);
         RouteEntity routeEntity = mapper.mapRouteToRouteEntity(element);
         routeDAO.updateElement(routeEntity);
     }
