+++ src/main/java/controller/service/impl/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package controller.service.impl;
 
+import controller.exception.InvalidDataRuntimeException;
 import controller.exception.ServiceLayerRuntimeException;
 import controller.exception.UserNotExistRuntimeException;
 import controller.service.UserService;
@@ -12,6 +13,7 @@ import org.apache.log4j.Logger;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class UserServiceImpl implements UserService {
@@ -29,8 +31,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findUserByLoginData(User user) throws ServiceLayerRuntimeException {
-        String encodedPassword = encoderPassword.encode(user.getPassword());
         LOGGER.info("Try to find user by login data");
+        if (Objects.isNull(user)) {
+            LOGGER.error("Incorrect findUserByLoginData value");
+            throw new InvalidDataRuntimeException("Incorrect findUserByLoginData value");
+        }
+        String encodedPassword = encoderPassword.encode(user.getPassword());
         UserEntity userEntity = mapper.mapUserToUserEntity(user);
         userEntity = userDAO.findByLogin(userEntity);
         if (!userEntity.getPassword().equals(encodedPassword)) {
@@ -44,6 +50,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public Integer insertElement(User element) {
         LOGGER.info("Inserting element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect insertElement value");
+            throw new InvalidDataRuntimeException("Incorrect insertElement value");
+        }
         UserEntity userEntity = mapper.mapUserToUserEntity(element);
         return userDAO.insertElement(userEntity);
     }
@@ -51,6 +61,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public User getElementById(Integer id) {
         LOGGER.info("Try to get element by id");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect getElementById value");
+            throw new InvalidDataRuntimeException("Incorrect getElementById value");
+        }
         UserEntity userEntity = userDAO.getElementById(id);
         return mapper.mapUserEntityToUser(userEntity);
     }
@@ -58,12 +72,20 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteElement(Integer id) {
         LOGGER.info("Deleting element");
+        if (Objects.isNull(id)) {
+            LOGGER.error("Incorrect deleteElement value");
+            throw new InvalidDataRuntimeException("Incorrect deleteElement value");
+        }
         userDAO.deleteElement(id);
     }
 
     @Override
     public void updateElement(User element) {
         LOGGER.info("Updating element");
+        if (Objects.isNull(element)) {
+            LOGGER.error("Incorrect updateElement value");
+            throw new InvalidDataRuntimeException("Incorrect updateElement value");
+        }
         UserEntity scheduleEntity = mapper.mapUserToUserEntity(element);
         userDAO.updateElement(scheduleEntity);
     }
